+++ src/slash-commands/ruokaa-keskusta/components/buttonRow.ts
@@ -1,31 +1,10 @@
-import { ExcludeEnum, MessageActionRow } from "discord.js";
+import { MessageActionRow } from "discord.js";
 import { MessageButtonStyles } from "discord.js/typings/enums";
-import { createButton } from "../../../util";
-import { RestaurantButtons } from "../consts";
-
-const generateButtonId = (restaurant: keyof typeof RestaurantButtons) =>
-    `ruokaa2 ${restaurant}`;
-
-const addButton = (
-    buttonRow: MessageActionRow,
-    restaurant: keyof typeof RestaurantButtons,
-    style?: ExcludeEnum<typeof MessageButtonStyles, "LINK">
-) => {
-    buttonRow.addComponents(
-        createButton(
-            generateButtonId(restaurant),
-            RestaurantButtons[restaurant],
-            style ?? MessageButtonStyles.PRIMARY
-        )
-    );
-};
+import { addButton } from "../../../utils/ruokaa-utils/componentUtils";
 
 export const createButtonRow = async () => {
     const buttonRow = new MessageActionRow();
 
-    addButton(buttonRow, "yolo");
-    addButton(buttonRow, "laseri");
-    addButton(buttonRow, "lutBuffet");
     addButton(buttonRow, "keskusta", MessageButtonStyles.SECONDARY);
     addButton(buttonRow, "skip", MessageButtonStyles.DANGER);
