--- src/slash-commands/ruokaa-skinnarila/embed/keskustaEmbed.ts
@@ -1,11 +1,0 @@
-import { MessageEmbed } from "discord.js";
-
-export const createKeskustaEmbed = () => {
-    const embed = new MessageEmbed()
-        .setTitle(`Keskusta`)
-        .setDescription(
-            "Ruokaa mm. Rosson, Luckiesin, ja Lalon tarjoilemana :]"
-        );
-
-    return embed;
-};
