+++ src/slash-commands/ruokaa-skinnarila/embed/menus.ts
@@ -11,13 +11,6 @@ export const createMenuEmbeds = () => {
     const lutBuffetEmbed = createMenuEmbed(ssNames.lutBuffet).setURL(
         "https://fi.jamix.cloud/apps/menu/?anro=97383&k=1&mt=4"
     );
-    const rossoEmbed = createMenuEmbed(ssNames.rosso)
-        .setDescription(
-            "Rossossa saatavilla päivittäin myös muunmuassa salaattia & pitsua!"
-        )
-        .setURL(
-            "https://www.raflaamo.fi/fi/ravintola/lappeenranta/rosso-isokristiina-lappeenranta/menu/lounas"
-        );
 
-    return [laserEmbed, yoloEmbed, lutBuffetEmbed, rossoEmbed];
+    return [laserEmbed, yoloEmbed, lutBuffetEmbed];
 };
