+++ src/slash-commands/ruokaa-skinnarila/index.ts
@@ -1,14 +1,12 @@
 import { CommandInteraction } from "discord.js";
 import { ApplicationCommandTypes } from "discord.js/typings/enums";
 import { SlashCommandModule } from "../../types";
+import { launchPuppeteer } from "../../utils/ruokaa-utils/puppeteerUtils";
 import { createButtonRow } from "./components/buttonRow";
 import { createMenuEmbeds } from "./embed/menus";
 import { createMenuAttachments } from "./file/menuAttachments";
-import { createKeskustaEmbed } from "./embed/keskustaEmbed";
 import { getAalefClips } from "./menus/getAalefClips";
-import { launchPuppeteer } from "./puppeteerUtils";
 import { getLutBuffetClip } from "./menus/getLutBuffetClip";
-import { getRossoClips } from "./menus/getRossoClips";
 
 const command: SlashCommandModule = {
     data: {
@@ -29,7 +27,6 @@ const ruokaa = async (interaction: CommandInteraction) => {
 
         await getAalefClips(browser);
         await getLutBuffetClip(browser);
-        await getRossoClips(browser);
 
         await browser.close();
     } catch (error) {
@@ -41,13 +38,12 @@ const ruokaa = async (interaction: CommandInteraction) => {
     }
 
     try {
-        const keskustaEmbed = createKeskustaEmbed();
         const buttonRow = await createButtonRow();
         const menuEmbeds = createMenuEmbeds();
         const menuAttachments = createMenuAttachments();
 
         await interaction.editReply({
-            embeds: [...menuEmbeds, keskustaEmbed],
+            embeds: [...menuEmbeds],
             files: menuAttachments,
             components: [buttonRow],
         });
