+++ src/slash-commands/ruokaa-skinnarila/menus/getAalefClips.ts
@@ -1,7 +1,14 @@
 import { Browser, Page, ScreenshotClip } from "puppeteer";
+import {
+    getWeekday,
+    getNextFinnishDay,
+} from "../../../utils/ruokaa-utils/generalUtils";
+import {
+    navigateToPage,
+    clickButton,
+    screenShot,
+} from "../../../utils/ruokaa-utils/puppeteerUtils";
 import { ssNames } from "../consts";
-import { clickButton, navigateToPage, screenShot } from "../puppeteerUtils";
-import { getWeekday, getNextFinnishDay } from "../utils";
 
 const PageWidth = 800;
 const PageHeight = 2000;
