+++ src/utils/ruokaa-utils/consts.ts
@@ -1,0 +1,11 @@
+export const DayMap = {
+    Sunday: 0,
+    Monday: 1,
+    Tuesday: 2,
+    Wednesday: 3,
+    Thursday: 4,
+    Friday: 5,
+    Saturday: 6,
+};
+
+export const DayChangeHourUtc = 12;
