+++ src/utils/ruokaa-utils/generalUtils.ts
